pyproject.toml
[project]
name = "deepagentsspec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.2",
    "configparser>=7.2.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "json-repair>=0.55.0",
    "networkx>=3.6.1",
    "nltk>=3.9.2",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pipmaster>=1.0.10",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.1",
    "pypdf2>=3.0.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "rouge>=1.0.1",
    "setuptools>=80.9.0",
    "tenacity>=9.1.2",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.40.0",
    "xxhash>=3.6.0",
]
